% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_defined.R
\name{print_defined}
\alias{print_defined}
\title{Display variables that have been defined}
\usage{
print_defined(env = parent.frame())
}
\arguments{
\item{env}{An environment to look variables from}
}
\description{
Display variables that have been defined
}
\examples{
opts <- list()
opts$a <- 3
preint_defined(as.environment(opts))
}
